requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-wordpress-xmlrpc==2.3
lxml==4.9.3
python-dotenv==1.0.0
